% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_hfp.R
\name{cast_datetime_cols}
\alias{cast_datetime_cols}
\title{Cast UNIX milliseconds columns to date / timestamp}
\usage{
cast_datetime_cols(x, ts_cols = c("received_at", "tst"), date_cols = c("oday"))
}
\arguments{
\item{x}{A data frame (of raw HFP data)}

\item{ts_cols}{Chr names of columns to cast from ms to timestamp}

\item{date_cols}{Chr names of columns to cast from ms to date}
}
\value{
A data frame
}
\description{
Cast UNIX milliseconds columns to date / timestamp
}
