% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_hfp.R
\name{datetimes_as_character}
\alias{datetimes_as_character}
\title{Format dates and timestamps for csv output}
\usage{
datetimes_as_character(
  x,
  ts_format = "\%Y-\%m-\%dT\%H:\%M:\%OS3Z",
  date_format = "\%Y-\%m-\%d"
)
}
\arguments{
\item{x}{A data frame of HFP data}

\item{ts_format}{Datetime format string for POSIXct columns}

\item{date_format}{Date format string for Date columns}
}
\value{
A data frame
}
\description{
Format dates and timestamps for csv output
}
