% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure_storage.R
\name{download_blobs_by_date}
\alias{download_blobs_by_date}
\title{Download csv dumps by date from Azure storage}
\usage{
download_blobs_by_date(
  d,
  storage_url = Sys.getenv("AZURE_STORAGE_URL"),
  blob_prefix = "csv/VehiclePosition",
  result_prefix = "vp",
  target_dir = Sys.getenv("RAW_HFP_DIR"),
  verbose = TRUE
)
}
\arguments{
\item{d}{A Date or a string of format \code{YYYY-MM-DD}}

\item{storage_url}{A storage URL string (defaults to env var \code{AZURE_STORAGE_URL})}

\item{blob_prefix}{A prefix defining a subset of blobs (defaults to \code{csv/VehiclePosition})}

\item{result_prefix}{A prefix to be added to the result file names (defaults to \code{vp})}

\item{target_dir}{Path to the target dir of downloaded files (defaults to env var \code{RAW_HFP_DIR})}

\item{verbose}{If \code{TRUE}, warn about missing or locally existing datasets and report downloads}
}
\description{
HFP csv dumps are saved by date and hour in the HSL Azure blob storage.
This function tries to download all the available dumps from a given date.
If a csv file with the same name already exists in \code{target_dir},
it is not downloaded (and should be removed or renamed manually, if necessary).
}
