% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_hfp.R
\name{milliseconds_to_date}
\alias{milliseconds_to_date}
\title{Cast UNIX epoch milliseconds to UTC date}
\usage{
milliseconds_to_date(x)
}
\arguments{
\item{x}{A numeric vector of UTC milliseconds since 1970-01-01}
}
\value{
A Date object
}
\description{
Cast UNIX epoch milliseconds to UTC date
}
