% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_hfp.R
\name{milliseconds_to_timestamp}
\alias{milliseconds_to_timestamp}
\title{Cast UNIX epoch milliseconds to UTC datetime timestamp}
\usage{
milliseconds_to_timestamp(x)
}
\arguments{
\item{x}{A numeric vector of UTC milliseconds since 1970-01-01}
}
\value{
A \code{lubridate} date-time / POSIXct object
}
\description{
Cast UNIX epoch milliseconds to UTC datetime timestamp
}
