% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_hfp.R
\name{raw_hfp_col_names}
\alias{raw_hfp_col_names}
\title{Get default column names for raw HFP data}
\usage{
raw_hfp_col_names()
}
\value{
A character vector of column names
}
\description{
Returns a character vector of default column names of raw HFP event data
in storage (44 columns in total).
The original csv datasets do not have column names.
}
