% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_hfp.R
\name{raw_hfp_col_spec}
\alias{raw_hfp_col_spec}
\title{Get default column type specification for raw HFP data}
\usage{
raw_hfp_col_spec()
}
\description{
Returns a \href{https://readr.tidyverse.org/reference/cols.html}{\code{cols()}} specification
of default columns in raw HFP event data in storage (44 columns in total).
}
